Rework int tests, models
* Remove unused fields from update certificate item model.
* Write integration test for read certificate endpoint.
* Remove duplicate integration tests from controller integration tests.
