+++ src/test/java/uk/gov/companieshouse/certificates/orders/api/controller/CertificateItemsControllerReadEndpointIntegrationTest.java
@@ -1,5 +1,6 @@
 package uk.gov.companieshouse.certificates.orders.api.controller;
 
+import org.bson.Document;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -8,25 +9,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
+import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
-import uk.gov.companieshouse.certificates.orders.api.model.CertificateItem;
 import uk.gov.companieshouse.certificates.orders.api.model.CompanyProfileResource;
 import uk.gov.companieshouse.certificates.orders.api.repository.CertificateItemRepository;
 import uk.gov.companieshouse.certificates.orders.api.service.CompanyService;
 import uk.gov.companieshouse.certificates.orders.api.service.IdGeneratorService;
-import wiremock.org.eclipse.jetty.http.HttpStatus;
 
 import java.util.Optional;
 
-import static java.util.Objects.isNull;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.Is.is;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static uk.gov.companieshouse.certificates.orders.api.util.TestConstants.*;
@@ -34,9 +28,9 @@ import static uk.gov.companieshouse.certificates.orders.api.util.TestConstants.*
 @AutoConfigureMockMvc
 @SpringBootTest
 @ActiveProfiles("feature-flags-enabled")
-public class CertificateItemsControllerInitialEndpointIntegrationTest {
+public class CertificateItemsControllerReadEndpointIntegrationTest {
 
-    private static final String INITIAL_CERTIFICATE_URL = "/orderable/certificates/initial";
+    private static final String CERTIFICATES_URL = "/orderable/certificates/";
     private static final String EXPECTED_ITEM_ID = "CRT-123456-123456";
     private static final String TOKEN_PERMISSION_VALUE = "user_orders=%s";
 
@@ -55,6 +49,9 @@ public class CertificateItemsControllerInitialEndpointIntegrationTest {
     @Autowired
     private CertificateItemRepository repository;
 
+    @Autowired
+    private MongoTemplate mongoTemplate;
+
     @AfterEach
     void tearDown() {
         repository.findById(EXPECTED_ITEM_ID).ifPresent(repository::delete);
@@ -62,46 +59,20 @@ public class CertificateItemsControllerInitialEndpointIntegrationTest {
 
     @ParameterizedTest(name = "{0}")
     @MethodSource("uk.gov.companieshouse.certificates.orders.api.controller." +
-            "CertificateItemsControllerInitialEndpointTestData#initialEndpointTestData")
-    @DisplayName("Create initial certificate endpoint")
-    void testInitialEndpoint(JsonRequestFixture requestFixture) throws Exception {
-        Optional.ofNullable(requestFixture.getCompanyStatus())
-                .ifPresent(status -> when(companyProfileResource.getCompanyStatus()).thenReturn(status));
-        Optional.ofNullable(requestFixture.getCompanyType())
-                .ifPresent(type -> when(companyProfileResource.getCompanyType()).thenReturn(type.getCompanyType()));
-        if (isNull(requestFixture.getCompanyServiceException())) {
-            when(companyService.getCompanyProfile(any())).thenReturn(companyProfileResource);
-        } else {
-            when(companyService.getCompanyProfile(any())).thenThrow(requestFixture.getCompanyServiceException());
-        }
-        when(companyProfileResource.getCompanyName()).thenReturn("ACME LIMITED");
-        when(idGeneratorService.autoGenerateId()).thenReturn(EXPECTED_ITEM_ID);
-        mockMvc.perform(post(INITIAL_CERTIFICATE_URL)
+            "CertificateItemsControllerReadEndpointTestData#readEndpointTestData")
+    @DisplayName("Read certificate endpoint")
+    void testReadEndpoint(JsonRequestFixture requestFixture) throws Exception {
+        Optional.ofNullable(requestFixture.getSavedResource())
+                .ifPresent(resource -> mongoTemplate.insert(Document.parse(resource), "certificates"));
+        mockMvc.perform(get(CERTIFICATES_URL + EXPECTED_ITEM_ID)
                         .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
                         .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
                         .header(ERIC_IDENTITY_TYPE_HEADER_NAME, ERIC_IDENTITY_TYPE_OAUTH2_VALUE)
                         .header(ERIC_IDENTITY_HEADER_NAME, ERIC_IDENTITY_VALUE)
                         .header(ERIC_AUTHORISED_USER_HEADER_NAME, ERIC_AUTHORISED_USER_VALUE)
-                        .header(ERIC_AUTHORISED_TOKEN_PERMISSIONS_HEADER_NAME, String.format(TOKEN_PERMISSION_VALUE, "create"))
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(requestFixture.getRequestBody()))
+                        .header(ERIC_AUTHORISED_TOKEN_PERMISSIONS_HEADER_NAME, String.format(TOKEN_PERMISSION_VALUE, "read")))
                 .andExpect(status().is(requestFixture.getExpectedResponseCode()))
                 .andExpect(content().json(requestFixture.getExpectedResponseBody()))
                 .andDo(MockMvcResultHandlers.print());
-
-        if (!HttpStatus.isSuccess(requestFixture.getExpectedResponseCode())) {
-            assertItemWasNotSaved(EXPECTED_ITEM_ID);
-        }
-    }
-
-    /**
-     * Verifies that the item that could have been created by the create item POST request cannot in fact be retrieved
-     * from the database.
-     *
-     * @param expectedItemId the expected ID of the newly created item
-     */
-    private void assertItemWasNotSaved(final String expectedItemId) {
-        final Optional<CertificateItem> retrievedCertificateItem = repository.findById(expectedItemId);
-        assertThat(retrievedCertificateItem.isPresent(), is(false));
     }
 }
