+++ src/test/java/uk/gov/companieshouse/certificates/orders/api/interceptor/UserAuthorisationInterceptorIntegrationTest.java
@@ -1,0 +1,182 @@
+package uk.gov.companieshouse.certificates.orders.api.interceptor;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
+import uk.gov.companieshouse.certificates.orders.api.model.CertificateItem;
+import uk.gov.companieshouse.certificates.orders.api.repository.CertificateItemRepository;
+import uk.gov.companieshouse.certificates.orders.api.util.PatchMediaType;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static uk.gov.companieshouse.certificates.orders.api.util.TestConstants.*;
+import static uk.gov.companieshouse.certificates.orders.api.util.TestConstants.ERIC_AUTHORISED_TOKEN_PERMISSIONS_HEADER_NAME;
+
+@AutoConfigureMockMvc
+@SpringBootTest
+@ActiveProfiles("feature-flags-enabled")
+public class UserAuthorisationInterceptorIntegrationTest {
+
+    private static final String CERTIFICATES_URL = "/orderable/certificates/";
+    private static final String EXPECTED_ITEM_ID = "CRT-123456-123456";
+    private static final String TOKEN_PERMISSION_VALUE = "user_orders=%s";
+    private static final int QUANTITY = 1;
+    private static final String OTHER_USER = "otherUser";
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Autowired
+    private ObjectMapper mapper;
+
+    @Autowired
+    private CertificateItemRepository repository;
+
+    @AfterEach
+    void tearDown() {
+        repository.findById(EXPECTED_ITEM_ID).ifPresent(repository::delete);
+    }
+
+    @Test
+    @DisplayName("Fails to create certificate item with incorrect token permission")
+    void createCertificateItemUnauthorizedTokenPermission() throws Exception {
+        // When and Then
+        mockMvc.perform(post(CERTIFICATES_URL)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(ERIC_IDENTITY_TYPE_HEADER_NAME, ERIC_IDENTITY_TYPE_OAUTH2_VALUE)
+                        .header(ERIC_IDENTITY_HEADER_NAME, ERIC_IDENTITY_VALUE)
+                        .header(ERIC_AUTHORISED_USER_HEADER_NAME, ERIC_AUTHORISED_USER_VALUE)
+                        .header(ERIC_AUTHORISED_TOKEN_PERMISSIONS_HEADER_NAME, String.format(TOKEN_PERMISSION_VALUE, "read"))
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isUnauthorized());
+    }
+
+    // TODO: move to authorisation interceptor integration tests
+    @Test
+    @DisplayName("Return not found when a certificate item does not exist")
+    void getCertificateItemReturnsNotFound() throws Exception {
+        // When and then
+        mockMvc.perform(get(CERTIFICATES_URL + "CHS0")
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(ERIC_IDENTITY_TYPE_HEADER_NAME, ERIC_IDENTITY_TYPE_OAUTH2_VALUE)
+                        .header(ERIC_IDENTITY_HEADER_NAME, ERIC_IDENTITY_VALUE)
+                        .header(ERIC_AUTHORISED_USER_HEADER_NAME, ERIC_AUTHORISED_USER_VALUE)
+                        .header(ERIC_AUTHORISED_TOKEN_PERMISSIONS_HEADER_NAME, String.format(TOKEN_PERMISSION_VALUE, "read"))
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isNotFound())
+                .andDo(MockMvcResultHandlers.print());
+    }
+
+    // TODO: move to authorisation interceptor integration tests
+    @Test
+    @DisplayName("Return unauthorised if Eric headers are not present")
+    void getCertificateItemReturnsUnauthorisedWhenEricHeadersAreNotPresent() throws Exception {
+        // When and then
+        mockMvc.perform(get(CERTIFICATES_URL + EXPECTED_ITEM_ID)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isUnauthorized())
+                .andDo(MockMvcResultHandlers.print());
+    }
+
+    // TODO: move to authorisation interceptor integration tests
+    @Test
+    @DisplayName("Return unauthorised if the user has not created the certificate")
+    void getCertificateItemReturnsUnauthorisedIfUserDidNotCreateCertificate() throws Exception {
+        // Given
+        final CertificateItem newItem = new CertificateItem();
+        newItem.setId(EXPECTED_ITEM_ID);
+        newItem.setQuantity(QUANTITY);
+        newItem.setUserId(OTHER_USER);
+        repository.save(newItem);
+
+        // When and then
+        mockMvc.perform(get(CERTIFICATES_URL + EXPECTED_ITEM_ID)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(ERIC_IDENTITY_TYPE_HEADER_NAME, ERIC_IDENTITY_TYPE_OAUTH2_VALUE)
+                        .header(ERIC_IDENTITY_HEADER_NAME, ERIC_IDENTITY_VALUE)
+                        .header(ERIC_AUTHORISED_USER_HEADER_NAME, ERIC_AUTHORISED_USER_VALUE)
+                        .header(ERIC_AUTHORISED_TOKEN_PERMISSIONS_HEADER_NAME, String.format(TOKEN_PERMISSION_VALUE, "read"))
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isUnauthorized())
+                .andDo(MockMvcResultHandlers.print());
+    }
+
+    // TODO: move to authorisation interceptor integration tests
+    @Test
+    @DisplayName("Return unauthorised if the user does not have the right token permission")
+    void getCertificateItemReturnsUnauthorisedIfMissingTokenPermission() throws Exception {
+        // Given
+        final CertificateItem newItem = new CertificateItem();
+        newItem.setId(EXPECTED_ITEM_ID);
+        newItem.setQuantity(QUANTITY);
+        newItem.setUserId(OTHER_USER);
+        repository.save(newItem);
+
+        // When and then
+        mockMvc.perform(get(CERTIFICATES_URL + EXPECTED_ITEM_ID)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(ERIC_IDENTITY_TYPE_HEADER_NAME, ERIC_IDENTITY_TYPE_OAUTH2_VALUE)
+                        .header(ERIC_IDENTITY_HEADER_NAME, ERIC_IDENTITY_VALUE)
+                        .header(ERIC_AUTHORISED_USER_HEADER_NAME, ERIC_AUTHORISED_USER_VALUE)
+                        .header(ERIC_AUTHORISED_TOKEN_PERMISSIONS_HEADER_NAME, String.format(TOKEN_PERMISSION_VALUE, "other"))
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isUnauthorized())
+                .andDo(MockMvcResultHandlers.print());
+    }
+
+    // TODO: move to authorisation interceptor integration tests
+    @Test
+    @DisplayName("Fails to create certificate item with incorrect token permission")
+    void updateCertificateItemUnauthorizedTokenPermission() throws Exception {
+        // Given
+        final CertificateItem newItem = new CertificateItem();
+        newItem.setId(EXPECTED_ITEM_ID);
+        newItem.setQuantity(QUANTITY);
+        newItem.setUserId(OTHER_USER);
+        repository.save(newItem);
+
+        // When and then
+        mockMvc.perform(patch(CERTIFICATES_URL + EXPECTED_ITEM_ID)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(ERIC_IDENTITY_TYPE_HEADER_NAME, ERIC_IDENTITY_TYPE_OAUTH2_VALUE)
+                        .header(ERIC_IDENTITY_HEADER_NAME, ERIC_IDENTITY_VALUE)
+                        .header(ERIC_AUTHORISED_USER_HEADER_NAME, ERIC_AUTHORISED_USER_VALUE)
+                        .header(ERIC_AUTHORISED_TOKEN_PERMISSIONS_HEADER_NAME, String.format(TOKEN_PERMISSION_VALUE, "create"))
+                        .contentType(PatchMediaType.APPLICATION_MERGE_PATCH))
+                .andExpect(status().isUnauthorized())
+                .andDo(MockMvcResultHandlers.print());
+
+    }
+
+    // TODO: move to authorisation interceptor integration tests
+    @Test
+    @DisplayName("Reports failure to find certificate item")
+    void updateCertificateItemReportsFailureToFindItem() throws Exception {
+        // When and then
+        mockMvc.perform(patch(CERTIFICATES_URL + EXPECTED_ITEM_ID)
+                        .header(REQUEST_ID_HEADER_NAME, TOKEN_REQUEST_ID_VALUE)
+                        .header(ERIC_IDENTITY_TYPE_HEADER_NAME, ERIC_IDENTITY_TYPE_OAUTH2_VALUE)
+                        .header(ERIC_IDENTITY_HEADER_NAME, ERIC_IDENTITY_VALUE)
+                        .header(ERIC_AUTHORISED_USER_HEADER_NAME, ERIC_AUTHORISED_USER_VALUE)
+                        .header(ERIC_AUTHORISED_TOKEN_PERMISSIONS_HEADER_NAME, String.format(TOKEN_PERMISSION_VALUE, "update"))
+                        .contentType(PatchMediaType.APPLICATION_MERGE_PATCH))
+                .andExpect(status().isNotFound())
+                .andDo(MockMvcResultHandlers.print());
+    }
+}
