+++ src/test/java/uk/gov/companieshouse/certificates/orders/api/validator/PatchItemRequestValidatorTest.java
@@ -129,83 +129,6 @@ class PatchItemRequestValidatorTest {
     }
 
     @Test
-    @DisplayName("ID is read only")
-    void getValidationErrorsRejectsReadOnlyId() throws IOException {
-        itemUpdate.setId(TOKEN_STRING);
-        assertFieldMustBeNullErrorProduced(ApiErrors.raiseError(ApiErrors.ERR_CERTIFICATE_ID_SUPPLIED, "id: must be null"));
-    }
-
-    @Test
-    @DisplayName("Description is read only")
-    void getValidationErrorsRejectsReadOnlyDescription() throws IOException {
-        itemUpdate.setDescription(TOKEN_STRING);
-        assertFieldMustBeNullErrorProduced(ApiErrors.raiseError(ApiErrors.ERR_DESCRIPTION_SUPPLIED, "description: must be null"));
-    }
-
-    @Test
-    @DisplayName("Description identifier is read only")
-    void getValidationErrorsRejectsReadOnlyDescriptionIdentifier() throws IOException {
-        itemUpdate.setDescriptionIdentifier(TOKEN_STRING);
-        assertFieldMustBeNullErrorProduced(ApiErrors.raiseError(ApiErrors.ERR_DESCRIPTION_IDENTIFIER, "description_identifier: must be null"));
-    }
-
-    @Test
-    @DisplayName("Description values are read only")
-    void getValidationErrorsRejectsReadOnlyDescriptionValues() throws IOException {
-        itemUpdate.setDescriptionValues(TOKEN_VALUES);
-        assertFieldMustBeNullErrorProduced(ApiErrors.raiseError(ApiErrors.ERR_DESCRIPTION_VALUES_SUPPLIED, "description_values: must be null"));
-    }
-
-    @Test
-    @DisplayName("Item costs are read only")
-    void getValidationErrorsRejectsReadOnlyItemCosts() throws IOException {
-        itemUpdate.setItemCosts(TOKEN_ITEM_COSTS);
-        assertFieldMustBeNullErrorProduced(ApiErrors.raiseError(ApiErrors.ERR_ITEM_COSTS_SUPPLIED, "item_costs: must be null"));
-    }
-
-    @Test
-    @DisplayName("Kind is read only")
-    void getValidationErrorsRejectsReadOnlyKind() throws IOException {
-        itemUpdate.setKind(TOKEN_STRING);
-        assertFieldMustBeNullErrorProduced(ApiErrors.raiseError(ApiErrors.ERR_KIND_SUPPLIED_LOCATION, "kind: must be null"));
-    }
-
-    @Test
-    @DisplayName("Etag is read only")
-    void getValidationErrorsRejectsReadOnlyEtag() throws IOException {
-        itemUpdate.setEtag(TOKEN_STRING);
-        assertFieldMustBeNullErrorProduced(ApiErrors.raiseError(ApiErrors.ERR_ETAG_SUPPLIED, "etag: must be null"));
-    }
-
-    @Test
-    @DisplayName("Postal delivery is read only")
-    void getValidationErrorsRejectsReadOnlyPostalDelivery() throws IOException {
-        itemUpdate.setPostalDelivery(TOKEN_POSTAL_DELIVERY_VALUE);
-        assertFieldMustBeNullErrorProduced(ApiErrors.raiseError(ApiErrors.ERR_POSTAL_DELIVERY, "postal_delivery: must be null"));
-    }
-
-    @Test
-    @DisplayName("Multiple read only fields rejected")
-    void getValidationErrorsRejectsMultipleReadOnlyFields() throws IOException {
-        // Given
-        itemUpdate.setDescriptionValues(TOKEN_VALUES);
-        itemUpdate.setItemCosts(TOKEN_ITEM_COSTS);
-        itemUpdate.setKind(TOKEN_STRING);
-        itemUpdate.setEtag(TOKEN_STRING);
-        final JsonMergePatch patch = patchFactory.patchFromPojo(itemUpdate);
-
-        // When
-        final List<ApiError> errors = validatorUnderTest.getValidationErrors(patch);
-
-        // Then
-        assertThat(errors,
-                containsInAnyOrder(ApiErrors.raiseError(ApiErrors.ERR_DESCRIPTION_VALUES_SUPPLIED, "description_values: must be null"),
-                        ApiErrors.raiseError(ApiErrors.ERR_ITEM_COSTS_SUPPLIED, "item_costs: must be null"),
-                        ApiErrors.raiseError(ApiErrors.ERR_KIND_SUPPLIED_LOCATION, "kind: must be null"),
-                        ApiErrors.raiseError(ApiErrors.ERR_ETAG_SUPPLIED, "etag: must be null")));
-    }
-
-    @Test
     @DisplayName("Quantity must be greater than 0")
     void getValidationErrorsRejectsZeroQuantity() throws IOException {
         // Given
