+++ src/test/resources/integrationTestData/create/negative/item_options_error/request.json
@@ -5,7 +5,6 @@
     "administrators_details": {
       "include_basic_information": true
     },
-    "collection_location": "cardiff",
     "contact_number": "12345678901",
     "delivery_method": "collection",
     "delivery_timescale": "standard",
@@ -16,11 +15,10 @@
       "include_country_of_residence": true,
       "include_dob_type": "partial"
     },
-    "forename": "John",
     "general_partner_details": {
       "include_basic_information": true
     },
-    "include_email_copy": false,
+    "include_email_copy": true,
     "include_good_standing_information": false,
     "limited_partner_details": {
       "include_basic_information": true
@@ -42,8 +40,7 @@
     "registered_office_address_details": {
       "include_address_records_type": "current-previous-and-prior",
       "include_dates": false
-    },
-    "surname": "Smith"
+    }
   },
   "quantity": 1
 }
