+++ src/test/resources/integrationTestData/initial/negative/company_status_invalid/request.json
@@ -1,0 +1,3 @@
+{
+  "company_number": "12345678"
+}
