+++ src/test/resources/integrationTestData/initial/negative/company_status_invalid/response.json
@@ -1,0 +1,5 @@
+{
+  "errors": [{
+    "error": "company-status-invalid"
+  }]
+}
