+++ src/test/resources/integrationTestData/initial/negative/unknown_properties_set/request.json
@@ -1,0 +1,4 @@
+{
+  "company_number": "12345678",
+  "unknown_property": "unknown"
+}
