+++ src/test/resources/integrationTestData/initial/negative/unknown_properties_set/response.json
@@ -1,0 +1,5 @@
+{
+  "errors": [{
+    "error": "json-processing-error"
+  }]
+}
