+++ src/test/resources/integrationTestData/update/negative/delivery_timescale_invalid/request.json
@@ -1,0 +1,6 @@
+{
+  "item_options": {
+    "delivery_timescale": "unknown"
+  },
+  "quantity": 1
+}
