+++ src/test/resources/integrationTestData/update/negative/delivery_timescale_invalid/response.json
@@ -1,0 +1,5 @@
+{
+  "errors": [{
+    "error": "json-processing-error"
+  }]
+}
