+++ src/test/resources/integrationTestData/update/negative/invalid_field_value/response.json
@@ -1,5 +1,7 @@
 {
-  "errors": [{
-    "error": "quantity-error"
-  }]
+  "errors": [
+    {
+      "error": "quantity-error"
+    }
+  ]
 }
