+++ src/test/resources/integrationTestData/update/negative/unknown_properties_set/item.json
@@ -1,0 +1,43 @@
+{
+  "_id": "CRT-123456-123456",
+  "created_at" : ISODate("2022-01-01T12:00:00.000Z"),
+  "updated_at" : ISODate("2022-01-01T12:00:00.000Z"),
+  "data": {
+    "company_name": "ACME LIMITED",
+    "company_number": "12345678",
+    "customer_reference": "Test",
+    "item_options": {
+      "certificate_type": "incorporation-with-all-name-changes",
+      "collection_location": "cardiff",
+      "company_type": "llp",
+      "company_status": "active",
+      "contact_number": "12345678901",
+      "delivery_method": "collection",
+      "delivery_timescale": "standard",
+      "designated_member_details": {
+        "include_address": true,
+        "include_appointment_date": true,
+        "include_basic_information": true,
+        "include_country_of_residence": true,
+        "include_dob_type": "partial"
+      },
+      "forename": "John",
+      "include_email_copy": false,
+      "include_good_standing_information": false,
+      "member_details": {
+        "include_address": true,
+        "include_appointment_date": true,
+        "include_basic_information": true,
+        "include_country_of_residence": true,
+        "include_dob_type": "partial"
+      },
+      "registered_office_address_details": {
+        "include_address_records_type": "current-previous-and-prior",
+        "include_dates": false
+      },
+      "surname": "Smith"
+    },
+    "quantity": 1
+  },
+  "user_id": "Y2VkZWVlMzhlZWFjY2M4MzQ3MT"
+}
