+++ src/test/resources/integrationTestData/update/negative/unknown_properties_set/request.json
@@ -1,6 +1,7 @@
 {
   "company_number": "12345678",
+  "unknown_field": "unknown",
   "item_options": {
   },
-  "quantity": 0
+  "quantity": 1
 }
