+++ src/test/resources/integrationTestData/update/negative/unknown_properties_set/response.json
@@ -1,0 +1,7 @@
+{
+  "errors": [
+    {
+      "error": "json-processing-error"
+    }
+  ]
+}
